Fixes required for EKS.
Under EKS tbe `extension-apiserver-authentication` does not provide a `client-ca-file`. This might be fixed in the future under EKS by suppliying the correct flag to the API server. In the meantime, one can get the CA certificate by extracting it from `kubectl`'s current context.

Also, fix a typo in the README.md file.